Add product to database

+++ src/main/java/com/example/shopping_cart/model/Product.java
@@ -9,13 +9,13 @@ public class Product {
     private int id;
     private String name;
     private String category;
-    private int price;
+    private double price;
     private String image;
 
     public Product() {
     }
 
-    public Product(int id, String name, String category, int price, String image) {
+    public Product(int id, String name, String category, double price, String image) {
         this.id = id;
         this.name = name;
         this.category = category;
@@ -23,7 +23,7 @@ public class Product {
         this.image = image;
     }
 
-    public Product( String name, String category, int price, String image) {
+    public Product( String name, String category, double price, String image) {
         this.name = name;
         this.category = category;
         this.price = price;
